refactor(frontend): move document and workspace presentation into feature modules
