+++ frontend/src/components/WorkspaceLoader.tsx
@@ -1,21 +1,3 @@
 "use client";
 
-import { useUser } from "@/contexts/UserContext";
-import { useWorkspaceContext } from "@/contexts/WorkspaceContext";
-import { useWorkspacePolling } from "@/modules/workspaces";
-
-const WorkspaceLoader = () => {
-   const { isLoggedIn } = useUser();
-   const { setAvailableWorkspaces, setWorkspacesLoading } =
-      useWorkspaceContext();
-
-   useWorkspacePolling({
-      isLoggedIn,
-      setAvailableWorkspaces,
-      setWorkspacesLoading
-   });
-
-   return null;
-};
-
-export default WorkspaceLoader;
+export { default } from "@/modules/workspaces/presentation/WorkspaceLoader";
