+++ frontend/src/modules/documents/presentation/CreateTextDocumentButton.tsx
@@ -1,0 +1,17 @@
+"use client";
+import { useRouter } from "next/navigation";
+
+import { Button } from "@/components/ui/button";
+
+export default function CreateTextDocumentButton() {
+   const router = useRouter();
+
+   const navigateToCreateTextDocument = () => {
+      router.push("/document/new");
+   };
+   return (
+      <Button variant="outline" onClick={navigateToCreateTextDocument}>
+         Create text document
+      </Button>
+   );
+}
