+++ frontend/src/modules/documents/presentation/index.ts
@@ -1,1 +1,16 @@
+export { default as CreateTextDocumentButton } from "./CreateTextDocumentButton";
 export * from "./document-provider";
+export { default as DocumentChat } from "./DocumentChat";
+export { default as DocumentCreateDialog } from "./DocumentCreateDialog";
+export { default as DocumentData } from "./DocumentData";
+export { default as DocumentDetailsPage } from "./DocumentDetailsPage";
+export { default as DocumentEditable } from "./DocumentEditable";
+export { default as DocumentList } from "./DocumentList";
+export { default as DocumentPerspectives } from "./DocumentPerspectives";
+export { default as DocumentPreview } from "./DocumentPreview";
+export { default as DocumentPreviewDOCX } from "./DocumentPreviewDOCX";
+export { default as DocumentPreviewPDF } from "./DocumentPreviewPDF";
+export { default as DocumentTags } from "./DocumentTags";
+export { default as DocumentUpload } from "./DocumentUpload";
+export { default as DocumentVersions } from "./DocumentVersions";
+export { default as FloatingChat } from "./FloatingChat";
