+++ frontend/src/modules/workspaces/presentation/index.ts
@@ -1,1 +1,11 @@
+export { default as DeleteWorkspaceDialog } from "./DeleteWorkspaceDialog";
+export { default as EmptyWorkspace } from "./EmptyWorkspace";
+export { NavWorkspaces } from "./NavWorkspaces";
 export * from "./workspace-provider";
+export { default as WorkspaceDialog } from "./WorkspaceDialog";
+export { default as WorkspaceLoader } from "./WorkspaceLoader";
+export { default as WorkspaceMenu } from "./WorkspaceMenu";
+export { default as WorkspacePage } from "./WorkspacePage";
+export { default as WorkspaceTitle } from "./WorkspaceTitle";
+export { default as WorkspaceTour } from "./WorkspaceTour";
+export { default as WorkspaceTypeDialog } from "./WorkspaceTypeDialog";
